Eliminated LDIF read failure due to incorrect args
The ldbadd command for Sshpubkey schema specifies the sam.ldb file twice, which `ldbadd` treats as a second LDIF file, which prints an error to the output (because parsing the ldb file as LDIF fails). Also strictly speaking the options should be specified before the LDIF file(s) on the command line. The existing code worked, but this eliminates a spurious error from the output that caused me to waste time troubleshooting.