Allow specification of which interfaces to bind to
In cases where the container has multiple interfaces/multiple IPs, it may be necessary to specify which interfaces Samba should use. In my case (microk8s+multus) it was listening additionally on a k8s-internal IP address and _registering that IP address in DNS_, which clients on the LAN would not be able to reach. This change allows specification of an environment variable `BINDINTERFACES` that becomes the `interfaces` value in smb.conf. The default is `${HOSTIP} lo` which excluded any IPs other than the HOSTIP value.

This is a change from previous behavior, where all interfaces would be used, but this new behavior is probably desirable. The old behavior can be preserved by providing `false` explicitly in `BINDINTERFACES`.